+++ CHANGELOG.md
@@ -5,6 +5,12 @@ Change Log
 unreleased
 --------------------------
 
+### New Features
+
+- Add outline menu to side bar.
+    - [ ] Outline trigger
+    - [ ] menu item for outline view
+
 ### Improvements
 
 - When “Indent with Tab key” is on, reduce indent level of the current line with Shift+Tab even when nothing is selected.